feat: refactor sponsorship model and update related components
