+++ src/app/components/admin/dashboard/dashboard.component.html
@@ -166,9 +166,9 @@
     </div>
 
     <div class="mt-4">
-        <p-fieldset legend="Finders Stats" [toggleable]="true">
+        <p-fieldset legend="Finder Stats" [toggleable]="true">
             <p class="text-surface-400 text-md mb-4">
-                The following statistics are based on finders enhanced by explorers.
+                The following statistics are based on findings made by explorers.
             </p>
             <div class="flex flex-wrap gap-4">
                 <p-card class="flex-1" styleClass="border border-surface shadow-none">
